fix: corrected date helper and improved data processing log output
